+++ changelog.md
@@ -7,6 +7,7 @@
 - Add dependabot
 - Update requirements
 - Drop pyupgrade, autoflake, flake8 in favor of ruff
+- Refactor cover, add set_combined_position
 
 # Version 2023.3.0 (2023-03-01)